fix: Add tickets-module for ticket-functionality
