+++ src/api/index.ts
@@ -1,8 +1,3 @@
 export {CancelToken, isCancel, default as client} from './client';
 export {autocomplete, reverse} from './geocoder';
 export {default as searchTrip} from './trips';
-export {
-  listRecentFareContracts,
-  search as searchOffers,
-  reserve as reserveOffers,
-} from './fareContracts';
