+++ src/screens/Ticketing/Tickets/TicketsScrollView.tsx
@@ -2,13 +2,12 @@ import React from 'react';
 import {View, RefreshControl} from 'react-native';
 import {ScrollView} from 'react-native-gesture-handler';
 import {NavigationProp, useNavigation} from '@react-navigation/native';
-import {FareContract} from '../../../api/fareContracts';
 import {StyleSheet, useTheme} from '../../../theme';
 import LinearGradient from 'react-native-linear-gradient';
 import SimpleTicket from '../Ticket';
 import hexToRgba from 'hex-to-rgba';
 import ThemeText from '../../../components/text';
-import {ActiveReservation} from '../../../TicketContext';
+import {ActiveReservation, FareContract} from '../../../tickets';
 import TicketReservation from './TicketReservation';
 import {RootStackParamList} from '../../../navigation';
 import ErrorBoundary from '../../../error-boundary';
