+++ src/tickets/index.ts
@@ -1,0 +1,8 @@
+export {
+  default as TicketContextProvider,
+  useTicketState,
+} from './TicketContext';
+
+export * from './api';
+export * from './types';
+export * from './utils';
